refactor(frontend): ESLint errors 2 件解消 (this aliasing / fast-refresh exports)
フロントエンドリファクタリング 7 連 PR の #2。eslint . が報告する 2 件の
error を構造的に解消する。warning (12 件) は別 PR で対応。

修正 1: SearchReplaceExtension.ts の @typescript-eslint/no-this-alias
  問題: addProseMirrorPlugins() 内で `const extensionThis = this`
        と this を local 変数に alias していた。typescript-eslint 公式推奨に違反。
  解決: ProseMirror plugin の state.apply() コールバックは extension 文脈外で呼ばれるが、
        必要なのは storage オブジェクト参照のみなので
        `const storage = this.storage as SearchReplaceState` を addProseMirrorPlugins
        の頭で取得し closure に閉じ込める。これで apply 内の参照が `storage.xxx` に
        統一され this aliasing が不要になる。

修正 2: useToast.tsx の react-refresh/only-export-components
  問題: 同一ファイルから ToastProvider (component) と useToast (hook) を export しており、
        Vite React Fast Refresh の HMR 制約に抵触。
  解決: 3 ファイルに分離 (Vite 公式推奨パターン):
    - hooks/useToastContext.ts (新規): React Context + ToastItem / ToastContextValue 型
    - components/ToastProvider.tsx (新規): provider component のみ
    - hooks/useToast.tsx: useContext で context を読む hook のみ
  関連 import 変更:
    - App.tsx: ToastProvider を ../components/ToastProvider から import
    - components/layout/__tests__/AppShell.test.tsx
    - hooks/__tests__/useToast.test.tsx
    - pages/__tests__/LoginCallback.test.tsx

検証:
- eslint . → 0 errors (旧 2 errors)、warning 12 は維持 (次 PR で対応)
- vitest run → 293 files / 2361 tests 全 pass
